move out domains code, flip dimensions of pts, try correct uniform sampling in a ball
