+++ examples/namedpoly/jump.jl
@@ -3,181 +3,7 @@ Copyright 2018, Chris Coey and contributors
 
 see description in examples/namedpoly/native.jl
 =#
-
-using Hypatia
-using MathOptInterface
-MOI = MathOptInterface
-using MultivariatePolynomials
-using DynamicPolynomials
-using SemialgebraicSets
-using JuMP
-using PolyJuMP
-using SumOfSquares
-using LinearAlgebra
-using Test
-
-import Combinatorics
-
-abstract type Domain end
-
-mutable struct Box <: Domain
-    l::Vector{Float64}
-    u::Vector{Float64}
-    function Box(l::Vector{Float64}, u::Vector{Float64})
-        @assert length(l) == length(u)
-        d = new()
-        d.l = l
-        d.u = u
-        return d
-    end
-end
-function Box(l::Vector{T}, u::Vector{T}) where T <: Real
-    return Box(Float64.(l), Float64.(u))
-end
-
-
-# (x-c)'Q(x-c) \leq 1
-# struct Ellipsoid <: Domain
-#     c::Vector{Float64}
-#     Q::Matrix{Float64}
-#     function Ellipsoid(c::Vector{Float64}, Q::Matrix{Float64})
-#         @assert isposdef(Q)
-#         @assert length(c) == size(Q, 1)
-#         d = new()
-#         d.c = c
-#         d.Q = Q
-#         return d
-#     end
-# end
-# function Ball(c::Vector{Float64})
-#     dim = length(c)
-#     return Ellipsoid(c, Matrix{Float64}(I, dim, dim))
-# end
-
-# should be an ellipsoid
-mutable struct Ball <: Domain
-    c::Vector{Float64}
-    r::Float64
-    function Ball(c::Vector{Float64}, r::Float64)
-        d = new()
-        d.c = c
-        d.r = r
-        return d
-    end
-end
-
-dimension(d::Box) = length(d.l)
-dimension(d::Ball) = length(d.c)
-
-function sample(d::Box, npts::Int)
-    dim = dimension(d)
-    pts = rand(dim, npts)
-    return (d.u + d.l)/2.0 .+ (d.u - d.l) .* (pts .- 0.5)
-end
-
-function sample(d::Ball, npts::Int)
-    # sample(Box(d.c), npts::Int)
-    n = length(d.c)
-    dim = dimension(d)
-    pts = 0.5 .- rand(dim, npts)
-    for i in 1:npts
-        if norm(pts[:, i]) > 0.5
-            pts[:, i] .*= 0.4999 / norm(pts[:, i]) #* sqrt(n))
-        end
-        pts[:, i] .= pts[:, i] * 2 * d.r + d.c
-    end
-    return pts
-end
-
-# struct BoxSurf <: Domain
-#     l::Vector{Float64}
-#     u::Vector{Float64}
-# end
-#
-# dimension(d::Box) = length(d.l)
-#
-#
-# struct BallSurf <: Domain
-#     c::Vector{Float64}
-#     r::Float64
-# end
-#
-# struct Ellipse <: Domain
-#     c::Vector{Float64}
-#     Q::Matrix{Float64}
-# end
-#
-# struct EllipseSurf <: Domain
-#     c::Vector{Float64}
-#     Q::Matrix{Float64}
-# end
-
-
-get_bss(dom::Domain, x) = error("")
-function get_bss(dom::Box, x)
-    bss = BasicSemialgebraicSet{Float64,Polynomial{true,Float64}}()
-    for i in 1:dimension(dom)
-        addinequality!(bss, (-x[i] + dom.u[i]) * (x[i] - dom.l[i]))
-    end
-    return bss
-end
-function get_bss(dom::Ball, x)
-    return @set sum((x - dom.c).^2) <= dom.r^2
-end
-
-
-# function get_domain(dom::Ellipsoid, x)
-#     bss = BasicSemialgebraicSet{Float64,Polynomial{true,Float64}}()
-#     lhs = (x - dom.c)' * dom.Q * (x - dom.c)
-#     addinequality!(bss, 1 - lhs)
-#     return bss
-# end
-
-function get_weights(::Box, bss::BasicSemialgebraicSet{Float64,Polynomial{true,Float64}}, pts)
-    m = length(bss.p)
-    U = size(pts, 2)
-    g = Vector{Vector{Float64}}(undef, m)
-    for i in 1:m
-        g[i] = bss.p[i].(pts[i,:])
-        @assert all(g[i] .> -1e-6)
-    end
-    return g
-end
-function get_weights(::Ball, bss::BasicSemialgebraicSet{Float64,Polynomial{true,Float64}}, pts)
-    U = size(pts, 2)
-    @assert length(bss.p) == 1
-    # sub_func(j) = dom.r^2 - sum((dom.c - pts[:, j]).^2)
-    sub_func(j) = bss.p[1](pts[:, j])
-    g = [sub_func(j) for j in 1:U]
-    @show g
-    return [g]
-end
-
-function get_P(ipts, d::Int, U::Int)
-    (n, npts) = size(ipts)
-    u = Hypatia.calc_u(n, 2d, ipts')
-    m = Vector{Float64}(undef, U)
-    m[1] = 2^n
-    M = Matrix{Float64}(undef, npts, U)
-    M[:,1] .= 1.0
-
-    col = 1
-    for t in 1:2d
-        for xp in Combinatorics.multiexponents(n, t)
-            col += 1
-            if any(isodd, xp)
-                m[col] = 0.0
-            else
-                m[col] = m[1]/prod(1.0 - abs2(xp[j]) for j in 1:n)
-            end
-            @. @views M[:,col] = u[1][:,xp[1]+1]
-            for j in 2:n
-                @. @views M[:,col] *= u[j][:,xp[j]+1]
-            end
-        end
-    end
-    return M
-end
+include(joinpath(dirname(@__DIR__()), "domains.jl"))
 
 function build_JuMP_namedpoly_SDP(
     x,
@@ -211,7 +37,7 @@ function build_JuMP_namedpoly_WSOS(
 
     pts_factor = n
 
-    # toggle between sampling and current Hypatia method
+    # toggle between new sampling and current Hypatia method, this is just for debugging purposes
     sample_pts = true
 
     if sample_pts
@@ -222,7 +48,7 @@ function build_JuMP_namedpoly_WSOS(
         F = qr!(Mp, Val(true))
         keep_pnt = F.p[1:U]
 
-        pts = candidate_pts[:,keep_pnt] # subset of points indexed with the support of w
+        pts = candidate_pts[keep_pnt,:] # subset of points indexed with the support of w
         P0 = M[keep_pnt,1:L] # subset of polynomial evaluations up to total degree d
         P = Array(qr(P0).Q)
 
@@ -230,7 +56,6 @@ function build_JuMP_namedpoly_WSOS(
     else
         L = binomial(n+d, n)
         (L, U, pts, P0, P, w) = Hypatia.interpolate(n, d, calc_w=false)
-        pts = pts'
         pts .*= 2.0
     end
 
@@ -240,7 +65,7 @@ function build_JuMP_namedpoly_WSOS(
 
     g = get_weights(dom, bss, pts)
     @assert length(g) == length(bss.p)
-    PWts = [sqrt.(g[i]) .* P0sub for i in 1:length(bss.p)]
+    PWts = [sqrt.(gi) .* P0sub for gi in g]
 
     wsos_cone = WSOSPolyInterpCone(U, [P, PWts...])
 
@@ -253,7 +78,7 @@ function build_JuMP_namedpoly_WSOS(
     @objective(model, Max, a)
     @constraints(model, begin
         q - a .* ones(U) in wsos_cone
-        [i in 1:U], f(pts[:,i]) == q[i]
+        [i in 1:U], f(pts[i,:]) == q[i]
     end)
 
     return model
